requirements.txt
# Core dependencies
pandas>=2.0.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0

# Optional - for enhanced features
matplotlib>=3.7.0  # For visualization
seaborn>=0.12.0    # For better charts
tabulate>=0.9.0    # For pretty table output

# Web dashboard
flask>=3.0.0       # Web framework
flask-cors>=4.0.0  # CORS support
streamlit>=1.28.0  # Streamlit dashboard
plotly>=5.17.0     # Interactive charts